requirements.txt
langchain>=0.1.17
langchain-community>=0.0.21
openai
faiss-cpu
python-dotenv
faiss-cpu